requirements.txt
Flask==2.2.2
requests==2.28.1
playwright==1.42.0
pandas==2.1.0
python-dateutil==2.8.2
pybit==5.5.0
python-dotenv==1.0.0
numpy==1.24.3
python-binance==1.0.19
Werkzeug==2.3.3
ccxt
SQLAlchemy